SpaceWeb_doc/requirements.txt
selenium>=4.0.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.2
cryptography>=42.0.8
